トピックアップデートエラーの解消
フォーカス後のページ遷移時にフォーカス確認の条件分岐を追加
（＋自動保存時間表示の修正）
